mysql/json:split compare,serde,binary into independent files
