+++ src/util/codec/mysql/json/mod.rs
@@ -11,6 +11,32 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod json;
+// FIXME(shirly): remove following later
+#![allow(dead_code)]
+mod binary;
+mod comparison;
+mod serde;
 
-pub use self::json::{Json, JsonEncoder, JsonDecoder};
+pub use self::binary::{JsonEncoder, JsonDecoder};
+
+const ERR_CONVERT_FAILED: &str = "Can not covert from ";
+
+use std::collections::BTreeMap;
+
+/// Json implements type json used in tikv, it specifies the following
+/// implementations:
+/// 1. Serialize `json` values into binary representation, and reading values
+///  back from the binary representation.
+/// 2. Serialize `json` values into readable string representation, and reading
+/// values back from string representation.
+/// 3. sql functions like `JSON_TYPE`, etc
+#[derive(Clone, Debug)]
+pub enum Json {
+    Object(BTreeMap<String, Json>),
+    Array(Vec<Json>),
+    I64(i64),
+    Double(f64),
+    String(String),
+    Boolean(bool),
+    None,
+}
